Refactor storage accessible into smaller files.
